Reduced scope of node __setattr__ to apply only to BuildConfig
